chore: move search.testdata to datasets
