+++ test/spec/algoliasearch.helper/pendingSearch.js
@@ -5,7 +5,7 @@ var algoliaSearch = require('algoliasearch');
 var algoliasearchHelper = require('../../../index');
 
 test('When searchOnce with callback, hasPendingRequests is true', function(done) {
-  var testData = require('../search.testdata')();
+  var testData = require('../../datasets/SearchParameters/search.dataset')();
   var client = algoliaSearch('dsf', 'dsfdf');
 
   var triggerCb;
@@ -36,7 +36,7 @@ test('When searchOnce with callback, hasPendingRequests is true', function(done)
 });
 
 test('When searchOnce with promises, hasPendingRequests is true', function(done) {
-  var testData = require('../search.testdata')();
+  var testData = require('../../datasets/SearchParameters/search.dataset')();
   var client = algoliaSearch('dsf', 'dsfdf');
 
   var triggerCb;
@@ -67,7 +67,7 @@ test('When searchOnce with promises, hasPendingRequests is true', function(done)
 });
 
 test('When searchForFacetValues, hasPendingRequests is true', function(done) {
-  var testData = require('../search.testdata')();
+  var testData = require('../../datasets/SearchParameters/search.dataset')();
   var client = algoliaSearch('dsf', 'dsfdf');
 
   var triggerCb;
@@ -98,7 +98,7 @@ test('When searchForFacetValues, hasPendingRequests is true', function(done) {
 });
 
 test('When helper.search(), hasPendingRequests is true', function(done) {
-  var testData = require('../search.testdata')();
+  var testData = require('../../datasets/SearchParameters/search.dataset')();
   var client = algoliaSearch('dsf', 'dsfdf');
 
   var triggerCb;
@@ -131,13 +131,13 @@ test('When helper.search(), hasPendingRequests is true', function(done) {
 });
 
 test('When helper.search() and one request is discarded, hasPendingRequests is true unless all come back', function(done) {
-  var testData = require('../search.testdata');
+  var testData = require('../../datasets/SearchParameters/search.dataset')();
   var client = algoliaSearch('dsf', 'dsfdf');
 
   var triggerCbs = [];
   client.search = function() {
     return new Promise(function(resolve) {
-      triggerCbs.push(function() { resolve(testData().response); });
+      triggerCbs.push(function() { resolve(testData.response); });
     });
   };
