+++ test/spec/algoliasearch.helper/searchOnce.js
@@ -5,7 +5,7 @@ var SearchParameters = require('../../../src/SearchParameters');
 var algoliasearchHelper = require('../../../index');
 
 test('searchOnce should call the algolia client according to the number of refinements and call callback with no error and with results when no error', function(done) {
-  var testData = require('../search.testdata')();
+  var testData = require('../../datasets/SearchParameters/search.dataset')();
 
   var client = {
     search: jest.fn().mockImplementationOnce(function() {
