+++ test/spec/requestBuilder.js
@@ -4,7 +4,7 @@ var requestBuilder = require('../../src/requestBuilder.js');
 var getQueries = requestBuilder._getQueries;
 
 test('The request builder should set analytics to subsequent queries', function() {
-  var testData = require('./search.testdata.js')();
+  var testData = require('../datasets/SearchParameters/search.dataset')();
   var searchParams = testData.searchParams;
 
   searchParams.analytics = true;
@@ -16,7 +16,7 @@ test('The request builder should set analytics to subsequent queries', function(
 });
 
 test('The request builder should set clickAnalytics to subsequent queries', function() {
-  var testData = require('./search.testdata.js')();
+  var testData = require('../datasets/SearchParameters/search.dataset')();
   var searchParams = testData.searchParams;
 
   searchParams.clickAnalytics = true;
@@ -28,7 +28,7 @@ test('The request builder should set clickAnalytics to subsequent queries', func
 });
 
 test('The request builder should should force analytics to false on subsequent queries if not specified', function() {
-  var testData = require('./search.testdata.js')();
+  var testData = require('../datasets/SearchParameters/search.dataset')();
   var searchParams = testData.searchParams;
 
   var queries = getQueries(searchParams.index, searchParams);
@@ -38,7 +38,7 @@ test('The request builder should should force analytics to false on subsequent q
 });
 
 test('The request builder should should force clickAnalytics to false on subsequent queries if not specified', function() {
-  var testData = require('./search.testdata.js')();
+  var testData = require('../datasets/SearchParameters/search.dataset')();
   var searchParams = testData.searchParams;
 
   var queries = getQueries(searchParams.index, searchParams);
@@ -46,4 +46,3 @@ test('The request builder should should force clickAnalytics to false on subsequ
   expect(queries[0].params.clickAnalytics).toBe(undefined);
   expect(queries[1].params.clickAnalytics).toBe(false);
 });
-
